fix: column wrong defined when use  namingStrategy

+++ spec/test.helper.ts
@@ -30,6 +30,9 @@ export class TestHelper {
         const tables = [...metadata.tables];
         while (tables.length > 0) {
             const table = tables.shift()!;
+            if (!table.name) {
+                return;
+            }
             const entity = table.target as typeof BaseEntity;
 
             await entity.query(`DROP TABLE ${table.name}`).catch((error) => {
@@ -55,7 +58,7 @@ export class TestHelper {
         });
     }
 
-    static getTypeOrmPgsqlModule(entities: TypeOrmModuleOptions['entities']) {
+    static getTypeOrmPgsqlModule(entities: TypeOrmModuleOptions['entities'], namingStrategy?: TypeOrmModuleOptions['namingStrategy']) {
         return TypeOrmModule.forRoot({
             type: 'postgres',
             database: process.env.POSTGRESQL_DATABASE_NAME,
@@ -65,6 +68,7 @@ export class TestHelper {
             synchronize: true,
             logging: true,
             logger: 'file',
+            namingStrategy,
         });
     }
